requirements.txt
asgiref==3.3.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.3
django-bootstrap-datepicker==1.4.0
django-bootstrap-daterangepicker==1.0.6
django-bootstrap3==14.2.0
django-import-export==2.4.0
django-jet==1.0.8
et-xmlfile==1.0.1
importlib-metadata==1.7.0
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
pkg-resources==0.0.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
unicodecsv==0.14.1
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
